requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
sqlalchemy
python-dotenv
pydantic
